Modify Mario: - Mario's height changes depending on Mario's level - Change Mario's pivot to fix the bug that makes Mario fall through the ground when level up / down

+++ DoAn-SE102/FireBall.cpp
@@ -15,7 +15,7 @@ void CFireBall::Update(DWORD dt) {
 	x += vx * dt;
 	y += vy * dt;
 	float cx, cy;
-	float screenWidth, screenHeight;
+	int screenWidth, screenHeight;
 	CGame::GetInstance()->GetCamPos(cx, cy);
 	screenWidth = CGame::GetInstance()->GetBackBufferWidth();
 	screenHeight = CGame::GetInstance()->GetBackBufferHeight();
