+++ DoAn-SE102/HardBlock.cpp
@@ -2,7 +2,7 @@
 #include "Game.h"
 void CHardBlock::Render() {
 	float cX, cY; CGame::GetInstance()->GetCamPos(cX, cY);
-	float screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
+	int screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
 	if (x < cX - HARD_BLOCK_WIDTH || x > cX + screenWidth) return;
 	if (y < cY - HARD_BLOCK_HEIGHT || y > cY + screenHeight + HARD_BLOCK_HEIGHT) return;
 	CSprites::GetInstance()->Get(HARD_BLOCK_SPRITE_ID)->Draw(x + HARD_BLOCK_WIDTH * 0.5f, y + HARD_BLOCK_HEIGHT * 0.5f);
