+++ DoAn-SE102/MultiTilesBlock.cpp
@@ -7,7 +7,7 @@ void CMultiTilesBlock::Render() {
 	for (int w = 0; w < width; w++) {
 		for (int h = 0; h < height; h++) {
 			float cX, cY; CGame::GetInstance()->GetCamPos(cX, cY);
-			float screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
+			int screenWidth = CGame::GetInstance()->GetBackBufferWidth(), screenHeight = CGame::GetInstance()->GetBackBufferHeight();
 			float tempX = x + cellWidth / 2.0f + w * cellWidth, tempY = y + cellHeight / 2.0f + h * cellHeight;
 			if (tempX < cX - cellWidth * 0.5f || tempX > cX + screenWidth + cellWidth * 0.5f) continue;
 			if (tempY < cY - cellHeight * 0.5f || tempY > cY + screenHeight + cellHeight * 0.5f) continue;
